# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AISystemCreate": {
        "properties": {
          "jurisdictions": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Jurisdictions",
            "type": "array"
          },
          "lifecycle_status": {
            "default": "draft",
            "enum": [
              "draft",
              "trial",
              "active",
              "retired"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "next_review_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review At"
          },
          "owner_name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Owner Name",
            "type": "string"
          },
          "provider_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Name"
          },
          "provider_type": {
            "enum": [
              "internal",
              "third_party"
            ],
            "title": "Provider Type",
            "type": "string"
          },
          "purpose": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Purpose",
            "type": "string"
          },
          "release_status": {
            "default": "draft",
            "enum": [
              "approved",
              "review_required",
              "blocked",
              "draft",
              "insufficient_evidence"
            ],
            "title": "Release Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "purpose",
          "provider_type",
          "jurisdictions",
          "owner_name"
        ],
        "title": "AISystemCreate",
        "type": "object"
      },
      "AISystemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AISystemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "default": 1,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 25,
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "items"
        ],
        "title": "AISystemListResponse",
        "type": "object"
      },
      "AISystemResponse": {
        "properties": {
          "assessment_version": {
            "title": "Assessment Version",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "jurisdictions": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Jurisdictions",
            "type": "array"
          },
          "lifecycle_status": {
            "default": "draft",
            "enum": [
              "draft",
              "trial",
              "active",
              "retired"
            ],
            "title": "Lifecycle Status",
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "next_review_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review At"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "owner_name": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Owner Name",
            "type": "string"
          },
          "provider_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Name"
          },
          "provider_type": {
            "enum": [
              "internal",
              "third_party"
            ],
            "title": "Provider Type",
            "type": "string"
          },
          "purpose": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Purpose",
            "type": "string"
          },
          "release_status": {
            "default": "draft",
            "enum": [
              "approved",
              "review_required",
              "blocked",
              "draft",
              "insufficient_evidence"
            ],
            "title": "Release Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "purpose",
          "provider_type",
          "jurisdictions",
          "owner_name",
          "id",
          "organization_id",
          "assessment_version"
        ],
        "title": "AISystemResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Role": {
        "enum": [
          "admin",
          "responsible_ai",
          "hr_reviewer",
          "legal_reviewer",
          "model_developer",
          "auditor",
          "procurement"
        ],
        "title": "Role",
        "type": "string"
      },
      "SessionResponse": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "organization_id": {
            "title": "Organization Id",
            "type": "string"
          },
          "organization_name": {
            "title": "Organization Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "display_name",
          "organization_id",
          "organization_name",
          "role",
          "permissions"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Organization-scoped recruitment AI assurance API. Responses are risk evidence, not legal determinations.",
    "title": "FairHire AI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/ai-systems": {
      "get": {
        "operationId": "list_ai_systems_v1_ai_systems_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISystemListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ai Systems",
        "tags": [
          "registry"
        ]
      },
      "post": {
        "operationId": "create_ai_system_v1_ai_systems_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AISystemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISystemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ai System",
        "tags": [
          "registry"
        ]
      }
    },
    "/v1/health": {
      "get": {
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "platform"
        ]
      }
    },
    "/v1/session": {
      "get": {
        "operationId": "session_v1_session_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-Organization-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Organization-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Dev-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Session",
        "tags": [
          "identity"
        ]
      }
    }
  }
}